requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
flask-sock==0.6.0
flask-cors==4.0.0
Pillow>=9.5.0
numpy>=1.26.0
PyPDF2==3.0.1
scikit-learn>=1.4.0
google-generativeai==0.3.1
firebase-admin==6.2.0
sentence-transformers>=4.1.0
--only-binary :all: 